# docs/swagger.yaml
definitions:
  model.LoginRequest:
    properties:
      email:
        example: dgkwon90@naver.com
        maxLength: 255
        type: string
      password:
        example: test1234
        maxLength: 255
        minLength: 8
        type: string
    type: object
  model.LoginResponse:
    properties:
      accessToken:
        example: '{uuid}'
        type: string
      birthDate:
        example: "2000-01-01"
        type: string
      email:
        example: dgkwon90@naver.com
        maxLength: 255
        type: string
      gender:
        enum:
        - male
        - ' female'
        - ' other'
        example: male
        type: string
      id:
        example: '{uuid}'
        type: string
      name:
        example: 권댜군
        maxLength: 255
        type: string
      nickName:
        example: 늘상
        maxLength: 36
        type: string
      phoneNumber:
        example: "01012344321"
        maxLength: 20
        type: string
      qna:
        $ref: '#/definitions/model.Qna'
      refreshToken:
        example: '{uuid}'
        type: string
    type: object
  model.Qna:
    properties:
      answer:
        example: blue
        maxLength: 255
        type: string
      question:
        example: What is my favorite color?
        maxLength: 255
        type: string
    type: object
  model.RefreshResponse:
    properties:
      accessToken:
        example: '{uuid}'
        type: string
      refreshToken:
        example: '{uuid}'
        type: string
    type: object
  model.RegisterRequest:
    properties:
      birthDate:
        example: "2000-01-01"
        type: string
      email:
        example: test@clobot.com
        maxLength: 255
        type: string
      gender:
        enum:
        - male
        - ' female'
        - ' other'
        example: male
        type: string
      name:
        example: tester
        maxLength: 255
        type: string
      nickName:
        example: tester
        maxLength: 36
        type: string
      password:
        example: test1234
        maxLength: 255
        type: string
      phoneNumber:
        example: "01012344321"
        maxLength: 20
        type: string
      qna:
        $ref: '#/definitions/model.Qna'
    type: object
  model.User:
    properties:
      birth_date:
        example: "1990-07-29"
        type: string
      email:
        example: dgkwon90@naver.com
        maxLength: 255
        type: string
      gender:
        enum:
        - male
        - ' female'
        - ' other'
        example: male
        type: string
      name:
        example: 권대근
        maxLength: 255
        type: string
      nickName:
        example: dgkwon90
        maxLength: 36
        type: string
      password:
        example: test1234
        maxLength: 255
        type: string
      phoneNumber:
        example: "01012344321"
        maxLength: 20
        type: string
      qna:
        $ref: '#/definitions/model.Qna'
    type: object
  model.UserResponse:
    properties:
      birthDate:
        example: "1990-07-29"
        type: string
      created_at:
        example: "2023-02-27 17:03:20"
        type: string
      deleted_at:
        example: "2023-02-27 17:03:20"
        type: string
      email:
        example: dgkwon90@naver.com
        maxLength: 255
        type: string
      gender:
        enum:
        - male
        - ' female'
        - ' other'
        example: male
        type: string
      id:
        example: '{uuid}'
        type: string
      name:
        example: 권대근
        maxLength: 255
        type: string
      nickName:
        example: dgkwon90
        maxLength: 36
        type: string
      phoneNumber:
        example: "01012344321"
        maxLength: 20
        type: string
      qna:
        $ref: '#/definitions/model.Qna'
      updated_at:
        example: "2023-02-27 17:03:20"
        type: string
    type: object
info:
  contact:
    email: dgkwon90@gmail.com
    name: API Support
    url: https://github.com/neulsang
  description: This is a Test auth api server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AMS Fantastic Auth Swagger API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login.
      parameters:
      - description: Login infomation
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResponse'
      summary: Login.
      tags:
      - Auth
  /api/v1/auth/logout:
    get:
      consumes:
      - application/json
      description: Logout.
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Logout.
      tags:
      - Auth
  /api/v1/auth/refresh:
    get:
      consumes:
      - application/json
      description: Request a new access token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RefreshResponse'
      security:
      - ApiKeyAuth: []
      summary: Request a new access token.
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user.
      parameters:
      - description: users infomation
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
      summary: Create a new user.
      tags:
      - Auth
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get all exists users information.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.UserResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all exists users information.
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user information by given ID.
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      summary: Delete user information by given ID.
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user information by given ID.
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: Get user information by given ID.
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update user information.
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: string
      - description: users infomation
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Update user information.
      tags:
      - Users
  /api/v1/users/me:
    get:
      consumes:
      - application/json
      description: Get my user information..
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Get my user information.
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"